Update dependencies, fix clippy warnings, adapt to nix bitsize changes (#337)
* Bump serde from 1.0.149 to 1.0.151

Bumps [serde](https://github.com/serde-rs/serde) from 1.0.149 to 1.0.151.
- [Release notes](https://github.com/serde-rs/serde/releases)
- [Commits](https://github.com/serde-rs/serde/compare/v1.0.149...v1.0.151)

---
updated-dependencies:
- dependency-name: serde
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* Bump serde_json from 1.0.89 to 1.0.91

Bumps [serde_json](https://github.com/serde-rs/json) from 1.0.89 to 1.0.91.
- [Release notes](https://github.com/serde-rs/json/releases)
- [Commits](https://github.com/serde-rs/json/compare/v1.0.89...v1.0.91)

---
updated-dependencies:
- dependency-name: serde_json
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* Bump serde_derive from 1.0.149 to 1.0.151

Bumps [serde_derive](https://github.com/serde-rs/serde) from 1.0.149 to 1.0.151.
- [Release notes](https://github.com/serde-rs/serde/releases)
- [Commits](https://github.com/serde-rs/serde/compare/v1.0.149...v1.0.151)

---
updated-dependencies:
- dependency-name: serde_derive
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* Bump clap from 4.0.29 to 4.0.32

Bumps [clap](https://github.com/clap-rs/clap) from 4.0.29 to 4.0.32.
- [Release notes](https://github.com/clap-rs/clap/releases)
- [Changelog](https://github.com/clap-rs/clap/blob/master/CHANGELOG.md)
- [Commits](https://github.com/clap-rs/clap/compare/v4.0.29...v4.0.32)

---
updated-dependencies:
- dependency-name: clap
  dependency-type: direct:production
  update-type: version-update:semver-patch
...


* Fix new clippy warnings

* (cargo-release) version 1.2.11

* Fix differing bitsize issue

+++ src/filesystems.rs
@@ -1,3 +1,4 @@
+use nix::libc::c_ulong;
 use nix::sys::statvfs::statvfs;
 use std::fs;
 
@@ -11,10 +12,10 @@ pub fn get_filesystems() -> Vec<Filesystem> {
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Filesystem {
     pub mount_point: String,
-    pub size_bytes: u64,
-    pub available_bytes: u64,
-    pub inodes: u64,
-    pub available_inodes: u64,
+    pub size_bytes: c_ulong,
+    pub available_bytes: c_ulong,
+    pub inodes: c_ulong,
+    pub available_inodes: c_ulong,
 }
 
 fn get_mount_points() -> Vec<String> {
@@ -29,10 +30,10 @@ fn filesystem_from_mount_point(mount_point: &str) -> Option<Filesystem> {
     match statvfs(mount_point) {
         Ok(stat) => Some(Filesystem {
             mount_point: mount_point.to_string(),
-            size_bytes: (stat.block_size() as u64) * (stat.blocks() as u64),
-            available_bytes: (stat.block_size() as u64) * (stat.blocks_available() as u64),
-            inodes: stat.files() as u64,
-            available_inodes: stat.files_available() as u64,
+            size_bytes: stat.block_size() * stat.blocks(),
+            available_bytes: stat.block_size() * stat.blocks_available(),
+            inodes: stat.files(),
+            available_inodes: stat.files_available(),
         }),
         Err(_) => None,
     }
